+++ python/src/client.py
@@ -62,8 +62,12 @@ class Client:
         return self.cache['keys']
 
     def put(self, key, value):
-        args = [f'"{key}"', value]
+        args = [key, value]
         return self.send('json', 'put', args)
+
+    def put_property(self, node_id, context_id, profile, property, value):
+        key = f'cns/{self.system_id}/nodes/{node_id}/contexts/{context_id}/provider/{profile}/properties/{property}'
+        self.put(key, value)
 
     def send(self, format, cmd, args=[]):
         for arg in args: