+++ rust/src/client.rs
@@ -298,9 +298,23 @@ impl Client {
     }
 
     pub fn put(&mut self, key: &str, value: &str) -> Result<(), Error> {
-        let args = vec![format!("\"{key}\""), value.to_string()];
+        let args = vec![key.to_string(), value.to_string()];
         let _ = self.send(Format::Json, "put", &args)?;
         Ok(())
+    }
+
+    pub fn put_property(
+        &mut self,
+        node_id: &str,
+        context_id: &str,
+        profile: &str,
+        property: &str,
+        value: &str,
+    ) -> Result<(), Error> {
+        let system_id = system::get_system_id()?;
+        let key =
+            format!("cns/{system_id}/nodes/{node_id}/contexts/{context_id}/provider/{profile}/properties/{property}");
+        self.put(&key, value)
     }
 
     pub fn send(&mut self, format: Format, cmd: &str, args: &[String]) -> Result<u64, Error> {